Testing whether, given an OU process, the best way to infer it is to track the mean or instantaneously follow the changes.
Conclusions suggest the latter.

It is important however to distinguish between a parametrisation of time scale (tau) vs. one of decay factors (alpha/beta), the first one simply says that faster inference systems are bounded to perform better than slower ones up to numerical errors, the second one implies that a specific decay constant will give the best results when the other one is fixed to a certain value.